requirements.txt
configparser==7.2.0
gpsd-py3==0.3.0
joblib==1.4.2
numpy==1.24.4
paho-mqtt==2.1.0
pyclip==0.7.0
pyrtlsdr==0.3.0
pyserial==3.5
scikit-learn==1.2.2
scipy==1.15.2
threadpoolctl==3.6.0
